Parallel loader (#1)
* testing

* update workflow

* Fix guild fetch everytime on calling class.

* update routung to parallel load Loader & Component
