--- .github/workflows/dev.node.js.yml
@@ -1,37 +1,0 @@
-name: Build and Deploy
-
-on:
-  push:
-    branches-ignore:
-      - master
-
-jobs:
-  test-docker:
-    name: "Test and Build as Docker"
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Build Docker Image
-        #continue-on-error: true
-        uses: docker/build-push-action@v4
-        with:
-          context: .
-          push: false
-          tags: user/app:latest
-
-  test-app:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        node-version: [ 18.x, 19.x ] # , 20.x ]
-        # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
-    steps:
-      - uses: actions/checkout@v3
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v3
-        with:
-          node-version: ${{ matrix.node-version }}
-          cache: 'npm'
-      - run: yarn install
-      - run: yarn build
