--- src/routing/guild/[guildId]/Index.tsx
@@ -1,23 +1,0 @@
-import { FC } from "react";
-import {
-	json,
-	LoaderFunction
-}             from "react-router-dom";
-
-const loader : LoaderFunction = async( { request, params } ) => {
-	return json( {} );
-};
-
-const Component : FC = () => {
-
-	return (
-		<>
-
-		</>
-	);
-};
-
-export {
-	Component,
-	loader
-};
