+++ src/routing/guild/[guildId]/Layout.tsx
@@ -3,28 +3,16 @@ import {
 	useState
 }                                 from "react";
 import {
-	json,
-	LoaderFunction,
 	Outlet,
 	useLoaderData,
 	useParams
 }                                 from "react-router-dom";
 import { validateLoginWithGuild } from "@hooks/useAuth";
 import GuildContext               from "@context/GuildContext";
-import { LoaderGuild }            from "@app/types/routing";
 import TopSubbar                  from "@comp/dashboard/TopSubbar";
 import TopNavbar                  from "@comp/dashboard/TopNavbar";
 import LeftSidebar                from "@comp/dashboard/LeftSidebar";
 import { usePageTitle }           from "@kyri123/k-reactutils";
-import { tRPC_Guild }             from "@lib/tRPC";
-import { EChannelType }           from "@shared/Enum/EDiscord";
-import {
-	DiscordForumChannel,
-	DiscordRole,
-	DiscordTextChannel,
-	DiscordVoiceChannel
-}                                 from "@shared/types/discord";
-import { MO_Mod }                 from "@shared/types/MongoDB";
 import {
 	ForumChannelContext,
 	TextChannelContext,
@@ -32,51 +20,7 @@ import {
 }                                 from "@context/ChannelContext";
 import { ModContext }             from "@context/ModContext";
 import { RoleContext }            from "@context/RoleContext";
-
-interface LoaderData extends LoaderGuild {
-	textChannels : DiscordTextChannel[];
-	forumChannels : DiscordForumChannel[];
-	voiceChannels : DiscordVoiceChannel[];
-	mods : MO_Mod[];
-	roles : DiscordRole[];
-}
-
-const loader : LoaderFunction = async( { params } ) => {
-	const query = await validateLoginWithGuild( params.guildId || "" );
-
-	if ( !query.loggedIn || !query.guildData ) {
-		window.location.replace( "/error/401" );
-	}
-
-	const [ textChannelsResult, forumChannelsResult, voiceChannelsResult, modsResult, rolesResult ] = await Promise.all( [
-		tRPC_Guild.channels.oftype.query( {
-			guildId: query.guildData?.guildId!,
-			type: EChannelType.text
-		} ),
-		tRPC_Guild.channels.oftype.query( {
-			guildId: query.guildData?.guildId!,
-			type: EChannelType.forum
-		} ),
-		tRPC_Guild.channels.oftype.query( {
-			guildId: query.guildData?.guildId!,
-			type: EChannelType.voice
-		} ),
-		tRPC_Guild.modupdates.mods.query( {
-			guildId: query.guildData?.guildId!
-		} ),
-		tRPC_Guild.roles.getrole.query( {
-			guildId: query.guildData?.guildId!
-		} )
-	] );
-
-	const textChannels : any[] = textChannelsResult?.channels || [];
-	const forumChannels : any[] = forumChannelsResult?.channels || [];
-	const voiceChannels : any[] = forumChannelsResult?.channels || [];
-	const roles : any[] = rolesResult?.roles || [];
-	const mods : any[] = modsResult?.mods || [];
-
-	return json<LoaderData>( { ...query, textChannels, forumChannels, voiceChannels, mods, roles } );
-};
+import { GuildLayoutLoaderData }  from "@guild/LayoutLoader";
 
 const Component : FC = () => {
 	const { guildId } = useParams();
@@ -88,7 +32,7 @@ const Component : FC = () => {
 		forumChannels,
 		roles,
 		voiceChannels
-	} = useLoaderData() as LoaderData;
+	} = useLoaderData() as GuildLayoutLoaderData;
 	usePageTitle( `Kbot 2.0 - ${ guildData?.guildData.name || "Unkown" }` );
 
 	const [ guild, setGuild ] = useState( () => guildData! );
@@ -99,10 +43,6 @@ const Component : FC = () => {
 
 	const triggerGuildUpdate = async() => {
 		const query = await validateLoginWithGuild( guildId || "" );
-
-		if ( !query.loggedIn || !query.guildData ) {
-			window.location.replace( "/error/401" );
-		}
 
 		setGuild( query.guildData! );
 	};
@@ -141,6 +81,5 @@ const Component : FC = () => {
 };
 
 export {
-	Component,
-	loader
+	Component
 };
