+++ src/routing/guild/[guildId]/Page.tsx
@@ -1,0 +1,14 @@
+import { FC } from "react";
+
+const Component : FC = () => {
+
+	return (
+		<>
+
+		</>
+	);
+};
+
+export {
+	Component
+};
