+++ src/routing/guild/[guildId]/patreon/Page.tsx
@@ -1,10 +1,4 @@
 import { FC }                from "react";
-import {
-	json,
-	LoaderFunction,
-	useLoaderData,
-	useParams
-}                            from "react-router-dom";
 import { Tabs }              from "flowbite-react";
 import {
 	BiCog,
@@ -13,21 +7,7 @@ import {
 import PatreonSettingsEditor from "@comp/patreon/PatreonSettingsEditor";
 import PatreonReleaseInput   from "@comp/patreon/PatreonReleaseInput";
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-interface LoaderData {
-}
-
-const loader : LoaderFunction = async( { params } ) => {
-	const { guildId } = params;
-
-	return json<LoaderData>( {} );
-};
-
 const Component : FC = () => {
-	const { guildId } = useParams();
-	// eslint-disable-next-line no-empty-pattern
-	const {} = useLoaderData() as LoaderData;
-
 	return (
 		<>
 			<div
@@ -55,6 +35,5 @@ const Component : FC = () => {
 	);
 };
 export {
-	Component,
-	loader
+	Component
 };
