+++ src/routing/register/[authCode]/Page.tsx
@@ -4,48 +4,23 @@ import {
 	useContext,
 	useRef,
 	useState
-}                          from "react";
+}                             from "react";
 import {
-	json,
-	LoaderFunction,
 	useLoaderData,
 	useNavigate,
 	useParams
-}                          from "react-router-dom";
-import { validateLogin }   from "@hooks/useAuth";
-import { usePageTitle }    from "@kyri123/k-reactutils";
-import { TextInput }       from "flowbite-react";
-import { SlLogin }         from "react-icons/all";
-import LoadButton          from "@comp/LoadButton";
-import { LoaderDataBase }  from "@app/types/routing";
-import { fireSwalFromApi } from "@lib/sweetAlert";
-import AuthContext         from "@context/AuthContext";
+}                             from "react-router-dom";
+import { usePageTitle }       from "@kyri123/k-reactutils";
+import { TextInput }          from "flowbite-react";
+import { SlLogin }            from "react-icons/all";
+import LoadButton             from "@comp/LoadButton";
+import { fireSwalFromApi }    from "@lib/sweetAlert";
+import AuthContext            from "@context/AuthContext";
 import {
 	tRCP_handleError,
 	tRPC_Public
-}                          from "@lib/tRPC";
-import { EApiTokenType }   from "@shared/Enum/EApiMethods";
-
-interface LoaderData extends LoaderDataBase {
-	tokenValid : boolean;
-}
-
-const loader : LoaderFunction = async( { params } ) => {
-	const { authCode } = params;
-	const result = await validateLogin();
-
-	const Response = await tRPC_Public.checktoken.mutate( {
-		token: authCode!,
-		type: EApiTokenType.reg
-	} ).catch( tRCP_handleError );
-
-	const tokenValid = !!Response?.valid;
-	if ( !tokenValid ) {
-		window.location.replace( "/error/401" );
-	}
-
-	return json<LoaderData>( { tokenValid, ...result } );
-};
+}                             from "@lib/tRPC";
+import { RegisterLoaderData } from "@routing/register/[authCode]/Loader";
 
 const Component : FC = () => {
 	usePageTitle( `Kbot 2.0 - Register` );
@@ -54,7 +29,7 @@ const Component : FC = () => {
 	const { authCode } = useParams();
 	const [ isLoading, setIsLoading ] = useState( false );
 	const [ inputError, setInputError ] = useState( [ false, false, false ] );
-	const { tokenValid } = useLoaderData() as LoaderData;
+	const { tokenValid } = useLoaderData() as RegisterLoaderData;
 
 	const loginRef = useRef<HTMLInputElement>( null );
 	const passwordRef = useRef<HTMLInputElement>( null );
@@ -124,6 +99,5 @@ const Component : FC = () => {
 };
 
 export {
-	Component,
-	loader
+	Component
 };
